Add function to sort lists
